backend/requirements.txt
# FastAPI Backend Dependencies for pyPortMan

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25

# Environment Variables
python-dotenv==1.0.0

# Zerodha KiteConnect
kiteconnect==5.0.0

# Scheduler
apscheduler==3.10.4

# HTTP Client
httpx==0.26.0

# Data Processing
pandas==2.2.0

# Pydantic for validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Async support
asyncpg==0.29.0  # For PostgreSQL (optional)
aiosqlite==0.19.0  # For SQLite async support
